binance_render_package/requirements.txt
ccxt
pandas
ta
python-dotenv
schedule
requests